requirements/testing.txt
-r base.txt

# Selenium
selenium==3.141.0

# Pytest
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
mixer==6.1.3

# Pytest dependencies
coverage==4.5.3
atomicwrites==1.3.0
attrs==19.1.0
more-itertools==7.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
six==1.12.0

# Mixer dependencies
Faker==0.9.1
python-dateutil==2.8.0
text-unidecode==1.2